[TASK] Improve URL normalization and cache key generation
This change the cache key generation by trimming ending slash to avoid
duplicate cache entry. This can be breaking on website where the ending
slash is required. URL normalization try to not break when the relative
URI don't start with a slash. The current implementation can be broken in
the case of a relative URI found in a subdirectory of the website.